+++ tradingstrategy/pair.py
@@ -63,7 +63,7 @@ import enum
 import pprint
 import warnings
 from collections import Counter
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from types import NoneType
 from typing import Optional, Iterable, Dict, TypeAlias
 
@@ -407,22 +407,42 @@ class DEXPair:
     sell_volume_30d: Optional[float] = None
 
     #: Buy token tax for this trading pair.
-    #: See :ref:`token-tax` for details.
-    buy_tax: Optional[float] = None
+    #:
+    #: See
+    #:  - :ref:`token-tax` what it means
+    #:  - :py:func:`tradingstrategy.utils.token_extra_data.load_extra_metadata` how to load
+    buy_tax: Optional[Percent] = None
 
     #: Transfer token tax for this trading pair.
     #: See :ref:`token-tax` for details.
-    transfer_tax: Optional[float] = None
+    #:
+    #: **Legacy**. Not used.
+    #:
+    transfer_tax: Optional[Percent] = None
 
     #: Sell tax for this trading pair.
-    #: See :ref:`token-tax` for details.
-    sell_tax: Optional[float] = None
+    #:
+    #: See
+    #:  - :ref:`token-tax` what it means
+    #:  - :py:func:`tradingstrategy.utils.token_extra_data.load_extra_metadata` how to load
+    sell_tax: Optional[Percent] = None
 
     #: Exchange name.
     #:
     #: Not part of the datasets. Added during the instance construction.
     #:
     exchange_name: Optional[str] = None
+
+    #: Any extra/user supplied data.
+    #:
+    #: Can contain:
+    #: - `top_pair_data`; TopPairData instance
+    #:
+    #: See
+    #:  - :py:meth:`token_sniffer_data`
+    #:  - :py:func:`tradingstrategy.utils.token_extra_data.load_extra_metadata` how to load
+    #:
+    other_data: Optional[dict] = field(default_factory=dict)
 
     def __repr__(self):
         exchange_name = self.exchange_name if self.exchange_name else f"{self.exchange_id}"
@@ -527,6 +547,17 @@ class DEXPair:
         else:
             return self.token1_decimals
 
+    @property
+    def token_sniffer_data(self) -> dict | None:
+        """Get TokenSniffer metadata.
+
+        Must be separarely loaded. See :py:func:`tradingstrategy.utils.tax.load_tokensniffer_metadata`
+        """
+        top_pair_data = self.other_data.get("top_pair_data")
+        if top_pair_data:
+            return top_pair_data.token_sniffer_data
+        return None
+
     def is_tradeable(
             self,
             liquidity_threshold=None,
@@ -599,6 +630,7 @@ class DEXPair:
         hints = {
             "chain_id": pa.uint64(),
             "dex_type": pa.string(),
+            "other_data": NoneType,  # TODO: Broken. PyArrow cannot handle object-like data.
         }
 
         return create_pyarrow_schema_for_dataclass(cls, hints=hints)
@@ -747,10 +779,12 @@ class PandasPairUniverse:
 
     """
 
-    def __init__(self,
-                 df: pd.DataFrame,
-                 build_index=True,
-                 exchange_universe: Optional[ExchangeUniverse]=None):
+    def __init__(
+            self,
+            df: pd.DataFrame,
+            build_index=True,
+            exchange_universe: Optional[ExchangeUniverse]=None
+    ):
         """
         :param df:
             The source DataFrame that contains all DEXPair entries